requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pywebview==5.1

